+++ components/Reusable/Logo.tsx
@@ -9,7 +9,7 @@ const Logo = () => {
                 <a>
                     <Image
                         className="justify-center"
-                        src="/logo.svg"
+                        src="/logo/tiara-logo.png"
                         alt="Picture of the author"
                         width={220}
                         height={50}
