+++ enums/ApiRoutes.ts
@@ -1,6 +1,7 @@
 export enum ApiRoutes {
-  BASE_URL = "http://tiarazar.com/api",
-  CLIENT_CATEGORIES = "/categories"
+  BASE_URL = "https://tiarazar.com/api",
+  CLIENT_CATEGORIES = "/categories",
+  USER_LOGIN="/login/mobile"
 }
