+++ messages/common/en.json
@@ -154,29 +154,30 @@
     "login": "ورود",
     "loginToTiara": "ورود به تیــارا",
     "loginDesc": "برای ورود و استفاده از خدمات تیارا شماره و رمز خود را وارد کنید",
-    "logout": "Logout",
-    "register": "Register",
-    "submit": "Submit",
-    "username": "Username",
-    "name": "Name",
+    "logout": "خروج",
+    "register": "ثبت نام",
+    "submit": "ثبت",
+    "username": "نام کاریری",
+    "name": "نام",
     "shipping_address": "Shipping Address",
-    "phone": "Phone",
+    "phone": "شماره همراه",
+    "phone_placeholder": "شماره همراه خود را وارد کنید",
     "email_address": "ایمیل",
     "email_address_placeholder": "ایمیل خود را وارد کنید",
     "password": "رمزعبور",
     "password_placeholder": "رمزعبور خود را وارد کنید",
-    "remember_me": "Remember me?",
-    "forgot_password": "Forgot your password?",
-    "not_member": "Not a member?",
-    "go_back_to": "Go back to",
-    "register_desc": "Your personal data will be used to support your experience throughout this website, to manage access to your account, and for other purposes described in our Privacy Policy",
-    "already_member": "Already a member?",
-    "email_already_exists": "Email already exists",
-    "error_occurs": "An error occurs",
-    "incorrect_email_password": "Incorrect email or password",
-    "login_successful": "Login Successful",
-    "register_successful": "Register Successful",
-    "profile": "Profile"
+    "remember_me": "مر به خاطر بسپار",
+    "forgot_password": "فراموشی رمز عبور",
+    "not_member": "عضو نیستید؟",
+    "go_back_to": "برگشت",
+    "register_desc": "حریم خصوصی اطلاعات شما کاملا حفظ خواهد شد",
+    "already_member": "عضو هستید؟",
+    "email_already_exists": "این ایمیل قبلا ثبت شده است",
+    "error_occurs": "خطایی رخ داد",
+    "incorrect_email_password": "اطلاعات نامعتبر می باشد",
+    "login_successful": "باموفقیت وارد شدید",
+    "register_successful": "ثبت نام موفق",
+    "profile": "حساب کاربری"
   },
   "Search": {
     "home": "Home",
