+++ pages/index.tsx
@@ -1,197 +1,91 @@
-import React, { useState, useEffect } from "react";
-import { GetStaticProps } from "next";
-import Image from "next/image";
-import { useTranslations } from "next-intl";
-import axios from "axios";
-
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
+import React, {useState} from 'react';
+import Input from "../components/Input/Input";
 import Button from "../components/Buttons/Button";
-import Slideshow from "../components/HeroSection/Slideshow";
-import OverlayContainer from "../components/OverlayContainer/OverlayContainer";
-import Card from "../components/Card/Card";
-import TestiSlider from "../components/TestiSlider/TestiSlider";
-import { apiProductsType, itemType } from "../context/cart/cart-types";
-import LinkButton from "../components/Buttons/LinkButton";
+import {useAuth} from "../context/AuthContext";
+import {useTranslations} from "next-intl";
+import {GetStaticProps} from "next";
+import Logo from "../components/Reusable/Logo";
+import {useRouter} from "next/router";
 
-// /bg-img/ourshop.png
-import ourShop from "../public/bg-img/ourshop.png";
-import {ApiRoutes} from "../enums/ApiRoutes";
-
-type Props = {
-  products: itemType[];
-};
-
-const Home: React.FC<Props> = ({ products }) => {
-  const t = useTranslations("Index");
-  const [currentItems, setCurrentItems] = useState(products);
-  const [isFetching, setIsFetching] = useState(false);
-
-  // useEffect(() => {
-  //   if (!isFetching) return;
-  //   const fetchData = async () => {
-  //     const res = await axios.get(
-  //       `${process.env.NEXT_PUBLIC_PROD_BACKEND_URL}/api/v1/products?order_by=createdAt.desc&offset=${currentItems.length}&limit=10`
-  //     );
-  //     const fetchedProducts = res.data.data.map((product: apiProductsType) => ({
-  //       ...product,
-  //       img1: product.image1,
-  //       img2: product.image2,
-  //     }));
-  //     setCurrentItems((products) => [...products, ...fetchedProducts]);
-  //     setIsFetching(false);
-  //   };
-  //   fetchData();
-  // }, [isFetching, currentItems.length]);
-
-  const handleSeemore = async (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
+const Index = () => {
+  const router = useRouter()
+  const auth = useAuth();
+  const [phone, setPhone] = useState("");
+  const [password, setPassword] = useState("");
+  const t = useTranslations("LoginRegister");
+  const [errorMsg, setErrorMsg] = useState("");
+  const [successMsg, setSuccessMsg] = useState("");
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    setIsFetching(true);
+    const loginResponse = await auth.login!(phone, password);
+    if (loginResponse.success) {
+      // setSuccessMsg("login_successful");
+      router.push("/shop")
+    } else {
+      alert("")
+      setErrorMsg("incorrect_email_password");
+    }
   };
 
   return (
-    <>
-      {/* ===== Header Section ===== */}
-      <Header />
+      <div className='p-4'>
+        <div className='flex justify-center mt-6 mb-12'>
+          <Logo/>
+        </div>
 
-      {/* ===== Carousel Section ===== */}
-      <Slideshow />
+        <h2 className="text-2xl text-right  font-bold leading-6 text-gray-900">{t("loginToTiara")}</h2>
+        <p className=" text-right mt-3 mb-8 font-medium leading-6 text-gray-900">{t("loginDesc")}</p>
 
-      <main id="main-content" className="-mt-20">
-        {/* ===== Category Section ===== */}
-        <section className="w-full h-auto py-10 border border-b-2 border-gray100">
-          <div className="app-max-width app-x-padding h-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4">
-            <div className="w-full sm:col-span-2 lg:col-span-2">
-              <OverlayContainer
-                imgSrc="/bg-img/banner_minipage1.jpg"
-                imgSrc2="/bg-img/banner_minipage1-tablet.jpg"
-                imgAlt="New Arrivals"
-              >
-                <LinkButton
-                  href="/product-category/new-arrivals"
-                  extraClass="absolute bottom-10-per sm:right-10-per z-20"
-                >
-                  {t("new_arrivals")}
-                </LinkButton>
-              </OverlayContainer>
-            </div>
-            <div className="w-full">
-              <OverlayContainer
-                imgSrc="/bg-img/banner_minipage2.jpg"
-                imgAlt="Women Collection"
-              >
-                <LinkButton
-                  href="/product-category/women"
-                  extraClass="absolute bottom-10-per z-20"
-                >
-                  {t("women_collection")}
-                </LinkButton>
-              </OverlayContainer>
-            </div>
-            <div className="w-full">
-              <OverlayContainer
-                imgSrc="/bg-img/banner_minipage3.jpg"
-                imgAlt="Men Collection"
-              >
-                <LinkButton
-                  href="/product-category/men"
-                  extraClass="absolute bottom-10-per z-20"
-                >
-                  {t("men_collection")}
-                </LinkButton>
-              </OverlayContainer>
-            </div>
-          </div>
-        </section>
+        <form onSubmit={handleSubmit} className="mt-2">
+          <label className='block text-right mb-2' htmlFor="">{t("phone")}</label>
+          <Input
+              type="phone"
+              placeholder={`${t("phone_placeholder")}`}
+              name="phone"
+              required
+              extraClass="w-full focus:border-gray500 text-right"
+              border="border-2 border-gray300 mb-4"
+              onChange={(e) => setPhone((e.target as HTMLInputElement).value)}
+              value={phone}
+          />
+          <label className='block text-right mb-2' htmlFor="">{t("password")}</label>
+          <Input
+              type="password"
+              placeholder={`${t("password_placeholder")}`}
+              name="password"
+              required
+              extraClass="w-full focus:border-gray500 mb-4 text-right"
+              border="border-2 border-gray300"
+              onChange={(e) => setPassword((e.target as HTMLInputElement).value)}
+              value={password}
+          />
+          {errorMsg !== "" && (
+              <div className="text-red text-sm mb-4 whitespace-nowrap">
+                {t(errorMsg)}
+              </div>
+          )}
 
-        {/* ===== Best Selling Section ===== */}
-        <section className="app-max-width w-full h-full flex flex-col justify-center mt-16 mb-20">
-          <div className="flex justify-center">
-            <div className="w-3/4 sm:w-1/2 md:w-1/3 text-center mb-8">
-              <h2 className="text-3xl mb-4">{t("best_selling")}</h2>
-              <span>{t("best_selling_desc")}</span>
-            </div>
-          </div>
-          <div className="grid grid-cols-2 md:grid-cols-4 gap-x-4 lg:gap-x-12 gap-y-6 mb-10 app-x-padding">
-            {/*<Card key={currentItems[1].id} item={currentItems[1]} />*/}
-            {/*<Card key={currentItems[2].id} item={currentItems[2]} />*/}
-            {/*<Card key={currentItems[3].id} item={currentItems[3]} />*/}
-            {/*<Card key={currentItems[4].id} item={currentItems[4]} />*/}
-          </div>
-        </section>
+          <Button
+              type="submit"
+              value={t("login")}
+              extraClass="w-full text-center text-xl mb-4"
+              size="lg"
+          />
+        </form>
+      </div>
 
-        {/* ===== Testimonial Section ===== */}
-        <section className="w-full hidden h-full py-16 md:flex flex-col items-center bg-lightgreen">
-          <h2 className="text-3xl">{t("testimonial")}</h2>
-          <TestiSlider />
-        </section>
-
-        {/* ===== Featured Products Section ===== */}
-        <section className="app-max-width app-x-padding my-16 flex flex-col">
-          <div className="text-center mb-6">
-            <h2 className="text-3xl">{t("featured_products")}</h2>
-          </div>
-          <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 gap-x-4 gap-y-10 sm:gap-y-6 mb-10">
-            {/*{currentItems.map((item) => (*/}
-            {/*  <Card key={item.id} item={item} />*/}
-            {/*))}*/}
-          </div>
-          <div className="flex justify-center">
-            <Button
-              value={!isFetching ? t("see_more") : t("loading")}
-              onClick={handleSeemore}
-            />
-          </div>
-        </section>
-
-        <div className="border-gray100 border-b-2"></div>
-
-        {/* ===== Our Shop Section */}
-        <section className="app-max-width mt-16 mb-20 flex flex-col justify-center items-center text-center">
-          <div className="textBox w-3/4 md:w-2/4 lg:w-2/5 mb-6">
-            <h2 className="text-3xl mb-6">{t("our_shop")}</h2>
-            <span className="w-full">{t("our_shop_desc")}</span>
-          </div>
-          <div className="w-full app-x-padding flex justify-center">
-            <Image src={ourShop} alt="Our Shop" />
-          </div>
-        </section>
-      </main>
-
-      {/* ===== Footer Section ===== */}
-      <Footer />
-    </>
   );
 };
 
-export const getStaticProps: GetStaticProps = async ({ locale }) => {
-//   let products: itemType[] = [];
-//   const res = await axios.get(
-//     `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/v1/products?order_by=createdAt.desc&limit=10`
-//   );
-//   const fetchedProducts = res.data;
-//   fetchedProducts.data.forEach((product: apiProductsType) => {
-//     products = [
-//       ...products,
-//       {
-//         id: product.id,
-//         name: product.name,
-//         price: product.price,
-//         img1: product.image1,
-//         img2: product.image2,
-//       },
-//     ];
-//   });
+
+export const getStaticProps: GetStaticProps = async ({locale}) => {
   return {
     props: {
       messages: {
         ...require(`../messages/common/${locale}.json`),
       },
-      // products,
-    }, // will be passed to the page component as props
+    },
   };
 };
 
-export default Home;
+export default Index;
